+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * java: update to java 15, please make sure to upgrade build/jenkins and all docker runtime to 14 before upgrading framework
     !!! recommended to use adoptopenjdk/openjdk15:alpine-jre for docker image
 * http: api client extends keep alive duration to 5 mins to improve internal api calls, track resolving dns and establishing connection performance 
+* http: set jdk networkaddress.cache.ttl/networkaddress.cache.negative.ttl to 300/0 by default
 
 ### 7.5.12 (09/23/2020 - 09/29/2020)
 * http: ForbiddenException supports errorCode, ipv4AccessControl throws forbiddenException with IP_ACCESS_DENIED error code